Move base message and utilities out of __init__.
