+++ generator/sbpg/targets/resources/sbp_construct_template.py.j2
@@ -16,7 +16,7 @@
 
 from construct import *
 import json
-from sbp import SBP
+from sbp.msg import SBP
 from sbp.utils import fmt_repr, exclude_fields, walk_json_dict, containerize
 import six
