+++ python/sbp/client/loggers/pickle_logger.py
@@ -7,7 +7,7 @@
 # THIS CODE AND INFORMATION IS PROVIDED "AS IS" WITHOUT WARRANTY OF ANY KIND,
 # EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED
 # WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.
-from ... import SBP
+from ...msg import SBP
 from .base_logger import BaseLogger, LogIterator
 import cPickle as pickle
