+++ python/sbp/file_io.py
@@ -26,7 +26,7 @@ device response.
 
 from construct import *
 import json
-from sbp import SBP
+from sbp.msg import SBP
 from sbp.utils import fmt_repr, exclude_fields, walk_json_dict, containerize
 import six
