+++ python/sbp/flash.py
@@ -23,7 +23,7 @@ are intended for internal-use only.
 
 from construct import *
 import json
-from sbp import SBP
+from sbp.msg import SBP
 from sbp.utils import fmt_repr, exclude_fields, walk_json_dict, containerize
 import six
