+++ python/sbp/logging.py
@@ -18,7 +18,7 @@ implementation-defined range (0x0000-0x00FF).
 
 from construct import *
 import json
-from sbp import SBP
+from sbp.msg import SBP
 from sbp.utils import fmt_repr, exclude_fields, walk_json_dict, containerize
 import six
