+++ python/sbp/navigation.py
@@ -27,7 +27,7 @@ and the RTK solution in tandem.
 
 from construct import *
 import json
-from sbp import SBP
+from sbp.msg import SBP
 from sbp.utils import fmt_repr, exclude_fields, walk_json_dict, containerize
 import six
