+++ python/sbp/observation.py
@@ -16,7 +16,7 @@ Satellite observation messages from the device.
 
 from construct import *
 import json
-from sbp import SBP
+from sbp.msg import SBP
 from sbp.utils import fmt_repr, exclude_fields, walk_json_dict, containerize
 import six
