+++ python/sbp/piksi.py
@@ -22,7 +22,7 @@ These messages are in the implementation-defined range
 
 from construct import *
 import json
-from sbp import SBP
+from sbp.msg import SBP
 from sbp.utils import fmt_repr, exclude_fields, walk_json_dict, containerize
 import six
