+++ python/sbp/settings.py
@@ -23,7 +23,7 @@ https://github.com/swift-nav/piksi\_firmware/blob/master/docs/settings.pdf
 
 from construct import *
 import json
-from sbp import SBP
+from sbp.msg import SBP
 from sbp.utils import fmt_repr, exclude_fields, walk_json_dict, containerize
 import six
