+++ python/sbp/tracking.py
@@ -17,7 +17,7 @@ Satellite code and carrier-phase tracking messages from the device.
 
 from construct import *
 import json
-from sbp import SBP
+from sbp.msg import SBP
 from sbp.utils import fmt_repr, exclude_fields, walk_json_dict, containerize
 import six
